// src/app/models/user.model.ts
export interface User_ {
  id: number;
  name?: any;
  show_name?: number;
  username?: any;
  email: string;
  email_verified_at?: any;
  role: number;
  phone?: any;
  country?: any;
  city?: any;
  address?: any;
  postal_code?: any;
  linkedin?: any;
  facebook?: any;
  youtube?: any;
  twitter?: any;
  instagram?: any;
  notifications: number;
  register_type: number;
  created_at: string;
  updated_at: string;
}